Consolidate audio tests into test_audio_ltx.py and fix two tests
  Merge the LTX audio tests into a single file. Fix test_audio_decode_perf,
  which skipped because _resolve_checkpoint() only honored LTX_CHECKPOINT
  when its basename matched a per-test allow-list — now the override is
  honored whenever the file exists. Fix test_audio_decode_e2e_psnr, which
  skipped when the pipeline lacked decode_audio_reference() — add a local
  host reference decode (mel-VAE + vocoder+BWE oracles loaded from the real
  checkpoint) so it runs its PSNR assertion.
